Remove release files; add github action/script to create releases/mix files
